+++ tests/Unit/Colors/Oklch/ColorTest.php
@@ -116,10 +116,10 @@ final class ColorTest extends BaseTestCase
     public function testNormalize(): void
     {
         $color = new Color(1, -0.4, 180);
-        $this->assertEquals([1.0, 0, .5, 1], $color->normalize());
+        $this->assertEquals([1.0, 0, .5, 1], $color->normalizedChannelValues());
 
         $color = new Color(1, 0.4, 90);
-        $this->assertEquals([1.0, 1.0, .25, 1], $color->normalize());
+        $this->assertEquals([1.0, 1.0, .25, 1], $color->normalizedChannelValues());
     }
 
     public function testToString(): void