+++ tests/Unit/Colors/Rgb/ColorTest.php
@@ -100,13 +100,13 @@ final class ColorTest extends BaseTestCase
     public function testNormalize(): void
     {
         $color = new Color(255, 0, 51);
-        $this->assertEquals([1.0, 0.0, 0.2, 1.0], $color->normalize());
+        $this->assertEquals([1.0, 0.0, 0.2, 1.0], $color->normalizedChannelValues());
 
         $color = new Color(255, 0, 51, 1);
-        $this->assertEquals([1.0, 0.0, 0.2, 1.0], $color->normalize());
+        $this->assertEquals([1.0, 0.0, 0.2, 1.0], $color->normalizedChannelValues());
 
         $color = new Color(255, 0, 51, .2);
-        $this->assertEquals([1.0, 0.0, 0.2, .2], $color->normalize());
+        $this->assertEquals([1.0, 0.0, 0.2, .2], $color->normalizedChannelValues());
     }
 
     public function testToString(): void